tests/wpt/tests/html/canvas/element/shadows/2d.shadow.alpha.5.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.shadow.alpha.5</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.shadow.alpha.5</h1>
<p class="desc">Shadows of shapes with alpha components are drawn correctly</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="2d.shadow.alpha.5.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Shadows of shapes with alpha components are drawn correctly");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = 'rgba(64, 0, 0, 0.5)';
  ctx.shadowColor = '#00f';
  ctx.shadowOffsetY = 50;
  ctx.fillRect(0, -50, 100, 50);

  _assertPixelApprox(canvas, 50,25, 127,0,127,255, 2);

});
</script>
